backend/pyproject.toml
[project]
name = "ui-validation-backend"
version = "0.1.0"
description = "Production FastAPI backend for UI change validation using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.1.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.0.0",
    "aiofiles>=24.1.0",
    "websockets>=14.0",
    # AI/ML Dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "ultralytics>=8.0.0",
    "accelerate>=0.20.0",
    "huggingface-hub>=0.15.0",
    "timm>=0.9.0",
    "openai>=1.0.0",
    "einops>=0.6.0",
    # Image Processing
    "pillow>=10.4.0",
    "opencv-python>=4.10.0",
    "scikit-image>=0.24.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    # Data Processing
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    # Background Tasks
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",
    # Monitoring & Logging
    "structlog>=24.1.0",
    "prometheus-client>=0.21.0",
    # Database (optional)
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
